requirements.txt
Flask>=3.0.0
Flask-SQLAlchemy>=3.1.0
Flask-CORS>=4.0.0
python-dotenv>=1.0.0
psycopg[binary]>=3.1.0
pytest>=8.0.0
